requirements_ci.txt
requests>=2.31
google-api-python-client>=2.100
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
gradio_client>=1.5.0
edge-tts>=6.1.9
